+++ CHANGELOG.md
@@ -5,5 +5,6 @@ All Notable changes to `drupal/opening-hours` module.
 ## [Unreleased]
 
 ### Added
+* DMOH-19: Added service configuration form.
 
 [Unreleased]: https://github.com/StadGent/drupal_module_opening-hours/compare/master...develop